part 4.21

+++ controllers/blogs.js
@@ -54,6 +54,13 @@ blogsRouter.put('/:id', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    const blog = await Blog.findById(req.params.id)
+    
+    if(!req.token || !decodedToken.id || blog.user.toString() !== decodedToken.id.toString()) {
+        return res.status(401).json({ error: 'token missing or invalid' })
+    }
+
     await Blog.findByIdAndRemove(req.params.id)
     res.status(204).end()
 })
